dropwizard-core/src/test/java/com/codahale/dropwizard/config/RequestLogConfigurationTest.java
package com.codahale.dropwizard.config;

import com.codahale.dropwizard.configuration.ConfigurationFactory;
import com.codahale.dropwizard.jackson.Jackson;
import com.codahale.dropwizard.logging.ConsoleLoggingOutput;
import com.codahale.dropwizard.logging.FileLoggingOutput;
import com.codahale.dropwizard.logging.SyslogLoggingOutput;
import com.fasterxml.jackson.databind.ObjectMapper;
import com.google.common.io.Resources;
import org.junit.Before;
import org.junit.Test;

import javax.validation.Validation;
import java.io.File;
import java.util.TimeZone;

import static org.fest.assertions.api.Assertions.assertThat;

public class RequestLogConfigurationTest {
    private RequestLogConfiguration requestLog;

    @Before
    public void setUp() throws Exception {
        final ObjectMapper objectMapper = Jackson.newObjectMapper();
        objectMapper.getSubtypeResolver().registerSubtypes(ConsoleLoggingOutput.class,
                                                           FileLoggingOutput.class,
                                                           SyslogLoggingOutput.class);
        this.requestLog = new ConfigurationFactory<>(RequestLogConfiguration.class,
                                                     Validation.buildDefaultValidatorFactory()
                                                               .getValidator(),
                                                     objectMapper)
                .build(new File(Resources.getResource("yaml/requestLog.yml").toURI()));
    }

    @Test
    public void defaultTimeZoneIsUTC() {
        assertThat(requestLog.getTimeZone())
            .isEqualTo(TimeZone.getTimeZone("UTC"));
    }
}
